Match episodes in the form s01.e01
Build episode ID by slicing the strings from the indexes of regexp
capture group matches. This isn't particularly clear, but it is the way
the rest of the function is written.

Of note, the expression had to be rewritten to make the capture groups
return the expected matches. It seems that a naive repeating group only
returns the start and end indexes of their last match.